+++ nall/string/cstring.hpp
@@ -7,11 +7,11 @@
 namespace nall {
 
 cstring::operator const char*() const { return data; }
-unsigned cstring::length() const { return strlen(data); }
+size_t cstring::length() const { return strlen(data); }
 bool cstring::operator==(const char *s) const { return !strcmp(data, s); }
 bool cstring::operator!=(const char *s) const { return  strcmp(data, s); }
-optional<unsigned> cstring::position (const char *key) const { return  strpos(data, key); }
-optional<unsigned> cstring::iposition(const char *key) const { return istrpos(data, key); }
+optional<size_t> cstring::position (const char *key) const { return  strpos(data, key); }
+optional<size_t> cstring::iposition(const char *key) const { return istrpos(data, key); }
 cstring& cstring::operator=(const char *data) { this->data = data; return *this; }
 cstring::cstring(const char *data) : data(data) {}
 cstring::cstring() : data("") {}
