+++ nall/string/wrapper.hpp
@@ -2,8 +2,8 @@
 
 namespace nall {
 
-unsigned string::length() const { return strlen(data); }
-unsigned string::capacity() const { return size; }
+size_t string::length() const { return strlen(data); }
+size_t string::capacity() const { return size; }
 
 template<unsigned limit> lstring string::split(const char *key) const { lstring result; result.split<limit>(key, data); return result; }
 template<unsigned limit> lstring string::isplit(const char *key) const { lstring result; result.isplit<limit>(key, data); return result; }
@@ -32,10 +32,10 @@ template<unsigned limit> string& string::ltrim(const char *key) { nall::ltrim<li
 template<unsigned limit> string& string::rtrim(const char *key) { nall::rtrim<limit>(data, key); return *this; }
 template<unsigned limit> string& string::trim(const char *key, const char *rkey) { nall::trim <limit>(data, key, rkey); return *this; }
 
-optional<unsigned> string::position(const char *key) const { return strpos(data, key); }
-optional<unsigned> string::iposition(const char *key) const { return istrpos(data, key); }
-optional<unsigned> string::qposition(const char *key) const { return qstrpos(data, key); }
-optional<unsigned> string::iqposition(const char *key) const { return iqstrpos(data, key); }
+optional<size_t> string::position(const char *key) const { return strpos(data, key); }
+optional<size_t> string::iposition(const char *key) const { return istrpos(data, key); }
+optional<size_t> string::qposition(const char *key) const { return qstrpos(data, key); }
+optional<size_t> string::iqposition(const char *key) const { return iqstrpos(data, key); }
 
 }
